+++ account-kit/smart-contracts/src/index.ts
@@ -122,3 +122,5 @@ export type * from "./ma-v2/account/common/modularAccountV2Base.js";
 export { createModularAccountV2 } from "./ma-v2/account/modularAccountV2.js";
 export type * from "./ma-v2/client/client.js";
 export { createModularAccountV2Client } from "./ma-v2/client/client.js";
+export type * from "./ma-v2/utils.js";
+export { getMAV2UpgradeToData } from "./ma-v2/utils.js";