+++ account-kit/smart-contracts/src/ma-v2/client/client.test.ts
@@ -3,7 +3,9 @@ import * as AACoreModule from "@aa-sdk/core";
 import {
   erc7677Middleware,
   LocalAccountSigner,
+  createSmartAccountClient,
   type SmartAccountSigner,
+  type UserOperationRequest_v7,
 } from "@aa-sdk/core";
 import {
   custom,
@@ -13,6 +15,11 @@ import {
   getContract,
   hashMessage,
   hashTypedData,
+  fromHex,
+  isAddress,
+  concat,
+  testActions,
+  type TestActions,
 } from "viem";
 import { HookType } from "../actions/common/types.js";
 import {
@@ -29,24 +36,42 @@ import {
   NativeTokenLimitModule,
   semiModularAccountBytecodeAbi,
 } from "@account-kit/smart-contracts/experimental";
-import { createModularAccountV2Client } from "@account-kit/smart-contracts";
+import {
+  createLightAccountClient,
+  createModularAccountV2Client,
+  type SignerEntity,
+} from "@account-kit/smart-contracts";
 import { local070Instance } from "~test/instances.js";
 import { setBalance } from "viem/actions";
 import { accounts } from "~test/constants.js";
 import { paymaster070 } from "~test/paymaster/paymaster070.js";
-import { alchemy, arbitrumSepolia } from "@account-kit/infra";
+import {
+  packAccountGasLimits,
+  packPaymasterData,
+} from "../../../../../aa-sdk/core/src/entrypoint/0.7.js";
+import { entryPoint07Abi } from "viem/account-abstraction";
+import {
+  alchemy,
+  arbitrumSepolia,
+  alchemyGasAndPaymasterAndDataMiddleware,
+} from "@account-kit/infra";
+import { getMAV2UpgradeToData } from "@account-kit/smart-contracts";
 
-// TODO: Include a snapshot to reset to in afterEach.
+// TODO: Include a snapshot to reset to in afterEach
 describe("MA v2 Tests", async () => {
   const instance = local070Instance;
 
   let client: ReturnType<typeof instance.getClient> &
-    ReturnType<typeof publicActions>;
+    ReturnType<typeof publicActions> &
+    TestActions;
 
   const isValidSigSuccess = "0x1626ba7e";
 
   beforeAll(async () => {
-    client = instance.getClient().extend(publicActions);
+    client = instance
+      .getClient()
+      .extend(publicActions)
+      .extend(testActions({ mode: "anvil" }));
   });
 
   const signer: SmartAccountSigner = new LocalAccountSigner(
@@ -366,7 +391,7 @@ describe("MA v2 Tests", async () => {
     let provider = (
       await givenConnectedProvider({
         signer,
-        usePaymaster: true,
+        paymasterMiddleware: "erc7677",
       })
     ).extend(installValidationActions);
 
@@ -375,31 +400,36 @@ describe("MA v2 Tests", async () => {
       value: parseEther("2"),
     });
 
-    const paymaster = paymaster070.getPaymasterStubData();
+    const paymaster = paymaster070.getPaymasterDetails().address;
 
     const hookInstallData = PaymasterGuardModule.encodeOnInstallData({
-      entityId: 0,
-      paymaster: "paymaster" in paymaster ? paymaster.paymaster : "0x0", // dummy value for paymaster address if it DNE
+      entityId: 1,
+      paymaster: paymaster,
     });
 
     const installResult = await provider.installValidation({
       validationConfig: {
-        moduleAddress: zeroAddress,
-        entityId: 0,
+        moduleAddress: getDefaultSingleSignerValidationModuleAddress(
+          provider.chain
+        ),
+        entityId: 1,
         isGlobal: true,
         isSignatureValidation: true,
         isUserOpValidation: true,
       },
       selectors: [],
-      installData: "0x",
+      installData: SingleSignerValidationModule.encodeOnInstallData({
+        entityId: 1,
+        signer: await sessionKey.getAddress(),
+      }),
       hooks: [
         {
           hookConfig: {
             address: getDefaultPaymasterGuardModuleAddress(provider.chain),
-            entityId: 0, // uint32
+            entityId: 1,
             hookType: HookType.VALIDATION,
             hasPreHooks: true,
-            hasPostHooks: true,
+            hasPostHooks: false,
           },
           initData: hookInstallData,
         },
@@ -409,8 +439,18 @@ describe("MA v2 Tests", async () => {
     // verify hook installtion succeeded
     await provider.waitForUserOperationTransaction(installResult);
 
+    // create session key client
+    const sessionKeyProvider = (
+      await givenConnectedProvider({
+        signer: sessionKey,
+        accountAddress: provider.account.address,
+        paymasterMiddleware: "erc7677",
+        signerEntity: { entityId: 1, isGlobalValidation: true },
+      })
+    ).extend(installValidationActions);
+
     // happy path: send a UO with correct paymaster
-    const result = await provider.sendUserOperation({
+    const result = await sessionKeyProvider.sendUserOperation({
       uo: {
         target: target,
         value: sendAmount,
@@ -419,16 +459,16 @@ describe("MA v2 Tests", async () => {
     });
 
     // verify if correct paymaster is used
-    const txnHash1 = provider.waitForUserOperationTransaction(result);
-    await expect(txnHash1).resolves.not.toThrowError();
+    const txnHash = sessionKeyProvider.waitForUserOperationTransaction(result);
+    await expect(txnHash).resolves.not.toThrowError();
 
     const hookUninstallData = PaymasterGuardModule.encodeOnUninstallData({
-      entityId: 0,
+      entityId: 1,
     });
 
     const uninstallResult = await provider.uninstallValidation({
       moduleAddress: zeroAddress,
-      entityId: 0,
+      entityId: 1,
       uninstallData: "0x",
       hookUninstallDatas: [hookUninstallData],
     });
@@ -443,7 +483,7 @@ describe("MA v2 Tests", async () => {
     let provider = (
       await givenConnectedProvider({
         signer,
-        usePaymaster: true,
+        paymasterMiddleware: "erc7677",
       })
     ).extend(installValidationActions);
 
@@ -452,31 +492,36 @@ describe("MA v2 Tests", async () => {
       value: parseEther("2"),
     });
 
-    const paymaster = paymaster070.getPaymasterStubData();
+    const paymaster = paymaster070.getPaymasterDetails().address;
 
     const hookInstallData = PaymasterGuardModule.encodeOnInstallData({
-      entityId: 0,
-      paymaster: "paymaster" in paymaster ? paymaster.paymaster : "0x0", // dummy value for paymaster address if it DNE
+      entityId: 1,
+      paymaster: paymaster,
     });
 
     const installResult = await provider.installValidation({
       validationConfig: {
-        moduleAddress: zeroAddress,
-        entityId: 0,
+        moduleAddress: getDefaultSingleSignerValidationModuleAddress(
+          provider.chain
+        ),
+        entityId: 1,
         isGlobal: true,
         isSignatureValidation: true,
         isUserOpValidation: true,
       },
       selectors: [],
-      installData: "0x",
+      installData: SingleSignerValidationModule.encodeOnInstallData({
+        entityId: 1,
+        signer: await sessionKey.getAddress(),
+      }),
       hooks: [
         {
           hookConfig: {
             address: getDefaultPaymasterGuardModuleAddress(provider.chain),
-            entityId: 0, // uint32
+            entityId: 1,
             hookType: HookType.VALIDATION,
             hasPreHooks: true,
-            hasPostHooks: true,
+            hasPostHooks: false,
           },
           initData: hookInstallData,
         },
@@ -487,13 +532,17 @@ describe("MA v2 Tests", async () => {
     await provider.waitForUserOperationTransaction(installResult);
 
     // sad path: send UO with no paymaster
-    let providerNoPaymaster = await givenConnectedProvider({
-      signer,
-      usePaymaster: false,
-    });
+    // create session key client
+    const sessionKeyProvider = (
+      await givenConnectedProvider({
+        signer: sessionKey,
+        accountAddress: provider.account.address,
+        signerEntity: { entityId: 1, isGlobalValidation: true },
+      })
+    ).extend(installValidationActions);
 
     await expect(
-      providerNoPaymaster.sendUserOperation({
+      sessionKeyProvider.sendUserOperation({
         uo: {
           target: target,
           value: sendAmount,
@@ -503,13 +552,17 @@ describe("MA v2 Tests", async () => {
     ).rejects.toThrowError();
 
     const hookUninstallData = PaymasterGuardModule.encodeOnUninstallData({
-      entityId: 0,
+      entityId: 1,
     });
 
     const uninstallResult = await provider.uninstallValidation({
-      moduleAddress: zeroAddress,
-      entityId: 0,
-      uninstallData: "0x",
+      moduleAddress: getDefaultSingleSignerValidationModuleAddress(
+        provider.chain
+      ),
+      entityId: 1,
+      uninstallData: SingleSignerValidationModule.encodeOnUninstallData({
+        entityId: 1,
+      }),
       hookUninstallDatas: [hookUninstallData],
     });
 
@@ -529,8 +582,52 @@ describe("MA v2 Tests", async () => {
       value: parseEther("2"),
     });
 
+    // install validation module
+    const installResult = await provider.installValidation({
+      validationConfig: {
+        moduleAddress: getDefaultSingleSignerValidationModuleAddress(
+          provider.chain
+        ),
+        entityId: 1,
+        isGlobal: true,
+        isSignatureValidation: true,
+        isUserOpValidation: true,
+      },
+      selectors: [],
+      installData: SingleSignerValidationModule.encodeOnInstallData({
+        entityId: 1,
+        signer: await sessionKey.getAddress(),
+      }),
+      hooks: [],
+    });
+
+    await provider.waitForUserOperationTransaction(installResult);
+
+    // create session key client
+    const sessionKeyProvider = (
+      await givenConnectedProvider({
+        signer: sessionKey,
+        accountAddress: provider.account.address,
+        signerEntity: { entityId: 1, isGlobalValidation: true },
+      })
+    ).extend(installValidationActions);
+
+    // verify we can call into the zero address before allow list hook is installed
+    const sendResultBeforeHookInstallation =
+      await sessionKeyProvider.sendUserOperation({
+        uo: {
+          target: zeroAddress,
+          value: 0n,
+          data: "0x",
+        },
+      });
+
+    await provider.waitForUserOperationTransaction(
+      sendResultBeforeHookInstallation
+    );
+
     const hookInstallData = AllowlistModule.encodeOnInstallData({
-      entityId: 0,
+      entityId: 1,
       inputs: [
         {
           target,
@@ -542,10 +639,13 @@ describe("MA v2 Tests", async () => {
       ],
     });
 
-    const installResult = await provider.installValidation({
+    // install hook
+    const installHookResult = await provider.installValidation({
       validationConfig: {
-        moduleAddress: zeroAddress,
-        entityId: 0,
+        moduleAddress: getDefaultSingleSignerValidationModuleAddress(
+          provider.chain
+        ),
+        entityId: 1,
         isGlobal: true,
         isSignatureValidation: true,
         isUserOpValidation: true,
@@ -556,7 +656,7 @@ describe("MA v2 Tests", async () => {
         {
           hookConfig: {
             address: getDefaultAllowlistModuleAddress(provider.chain),
-            entityId: 0, // uint32
+            entityId: 1,
             hookType: HookType.VALIDATION,
             hasPreHooks: true,
             hasPostHooks: false,
@@ -566,11 +666,11 @@ describe("MA v2 Tests", async () => {
       ],
     });
 
-    await provider.waitForUserOperationTransaction(installResult);
+    await provider.waitForUserOperationTransaction(installHookResult);
 
     // Test that the allowlist is active.
     // We should *only* be able to call into the target address, as it's the only address we passed to onInstall.
-    const sendResult = await provider.sendUserOperation({
+    const sendResult = await sessionKeyProvider.sendUserOperation({
       uo: {
         target: target,
         value: 0n,
@@ -582,7 +682,7 @@ describe("MA v2 Tests", async () => {
 
     // This should revert as we're calling an address separate fom the allowlisted target.
     await expect(
-      provider.sendUserOperation({
+      sessionKeyProvider.sendUserOperation({
         uo: {
           target: zeroAddress,
           value: 0n,
@@ -592,7 +692,7 @@ describe("MA v2 Tests", async () => {
     ).rejects.toThrowError();
 
     const hookUninstallData = AllowlistModule.encodeOnUninstallData({
-      entityId: 0,
+      entityId: 1,
       inputs: [
         {
           target,
@@ -605,24 +705,17 @@ describe("MA v2 Tests", async () => {
     });
 
     const uninstallResult = await provider.uninstallValidation({
-      moduleAddress: zeroAddress,
-      entityId: 0,
-      uninstallData: "0x",
+      moduleAddress: getDefaultSingleSignerValidationModuleAddress(
+        provider.chain
+      ),
+      entityId: 1,
+      uninstallData: SingleSignerValidationModule.encodeOnUninstallData({
+        entityId: 1,
+      }),
       hookUninstallDatas: [hookUninstallData],
     });
 
     await provider.waitForUserOperationTransaction(uninstallResult);
-
-    // Post-uninstallation, we should now be able to call into any address successfully.
-    const postUninstallSendResult = await provider.sendUserOperation({
-      uo: {
-        target: zeroAddress,
-        value: 0n,
-        data: "0x",
-      },
-    });
-
-    await provider.waitForUserOperationTransaction(postUninstallSendResult);
   });
 
   it("installs native token limit module, uses, then uninstalls", async () => {
@@ -637,16 +730,59 @@ describe("MA v2 Tests", async () => {
 
     const spendLimit = parseEther("0.5");
 
+    const installResult = await provider.installValidation({
+      validationConfig: {
+        moduleAddress: getDefaultSingleSignerValidationModuleAddress(
+          provider.chain
+        ),
+        entityId: 1,
+        isGlobal: true,
+        isSignatureValidation: true,
+        isUserOpValidation: true,
+      },
+      selectors: [],
+      installData: SingleSignerValidationModule.encodeOnInstallData({
+        entityId: 1,
+        signer: await sessionKey.getAddress(),
+      }),
+      hooks: [],
+    });
+
+    await provider.waitForUserOperationTransaction(installResult);
+
+    // create session key client
+    const sessionKeyProvider = (
+      await givenConnectedProvider({
+        signer: sessionKey,
+        accountAddress: provider.account.address,
+        signerEntity: { entityId: 1, isGlobalValidation: true },
+      })
+    ).extend(installValidationActions);
+
+    // Sending over the limit should currently pass
+    const preHookInstallationSendResult = await provider.sendUserOperation({
+      uo: {
+        target: target,
+        value: parseEther("0.6"),
+        data: "0x",
+      },
+    });
+    await provider.waitForUserOperationTransaction(
+      preHookInstallationSendResult
+    );
+
     // Let's verify the module's limit is set correctly after installation
     const hookInstallData = NativeTokenLimitModule.encodeOnInstallData({
-      entityId: 0,
+      entityId: 1,
       spendLimit,
     });
 
-    const installResult = await provider.installValidation({
+    const installHookResult = await provider.installValidation({
       validationConfig: {
-        moduleAddress: zeroAddress,
-        entityId: 0,
+        moduleAddress: getDefaultSingleSignerValidationModuleAddress(
+          provider.chain
+        ),
+        entityId: 1,
         isGlobal: true,
         isSignatureValidation: true,
         isUserOpValidation: true,
@@ -657,7 +793,7 @@ describe("MA v2 Tests", async () => {
         {
           hookConfig: {
             address: getDefaultNativeTokenLimitModuleAddress(provider.chain),
-            entityId: 0,
+            entityId: 1,
             hookType: HookType.VALIDATION,
             hasPreHooks: true,
             hasPostHooks: false,
@@ -667,7 +803,7 @@ describe("MA v2 Tests", async () => {
         {
           hookConfig: {
             address: getDefaultNativeTokenLimitModuleAddress(provider.chain),
-            entityId: 0,
+            entityId: 1,
             hookType: HookType.EXECUTION,
             hasPreHooks: true,
             hasPostHooks: false,
@@ -677,10 +813,10 @@ describe("MA v2 Tests", async () => {
       ],
     });
 
-    await provider.waitForUserOperationTransaction(installResult);
+    await provider.waitForUserOperationTransaction(installHookResult);
 
     // Try to send less than the limit - should pass
-    const passingSendResult = await provider.sendUserOperation({
+    const passingSendResult = await sessionKeyProvider.sendUserOperation({
       uo: {
         target: target,
         value: parseEther("0.05"), // below the 0.5 limit
@@ -691,7 +827,7 @@ describe("MA v2 Tests", async () => {
 
     // Try to send more than the limit - should fail
     await expect(
-      provider.sendUserOperation({
+      sessionKeyProvider.sendUserOperation({
         uo: {
           target: target,
           value: parseEther("0.6"), // passing the 0.5 limit
@@ -701,30 +837,24 @@ describe("MA v2 Tests", async () => {
     ).rejects.toThrowError();
 
     const hookUninstallData = NativeTokenLimitModule.encodeOnUninstallData({
-      entityId: 0,
+      entityId: 1,
     });
 
     const uninstallResult = await provider.uninstallValidation({
-      moduleAddress: zeroAddress,
-      entityId: 0,
-      uninstallData: "0x",
+      moduleAddress: getDefaultSingleSignerValidationModuleAddress(
+        provider.chain
+      ),
+      entityId: 1,
+      uninstallData: SingleSignerValidationModule.encodeOnUninstallData({
+        entityId: 1,
+      }),
       hookUninstallDatas: [hookUninstallData, "0x"],
     });
 
     await provider.waitForUserOperationTransaction(uninstallResult);
-
-    // Sending over the limit should now pass
-    const postUninstallSendResult = await provider.sendUserOperation({
-      uo: {
-        target: target,
-        value: parseEther("0.6"),
-        data: "0x",
-      },
-    });
-    await provider.waitForUserOperationTransaction(postUninstallSendResult);
   });
 
-  it("installs time range module, then uninstalls module within valid time range", async () => {
+  it("installs time range module, sends transaction within valid time range", async () => {
     let provider = (
       await givenConnectedProvider({
         signer,
@@ -736,92 +866,44 @@ describe("MA v2 Tests", async () => {
       value: parseEther("2"),
     });
 
-    const hookInstallData = TimeRangeModule.encodeOnInstallData({
-      entityId: 0,
-      validAfter: 0,
-      validUntil: 10000000000,
-    });
-
-    const installResult = await provider.installValidation({
-      validationConfig: {
-        moduleAddress: zeroAddress,
-        entityId: 0,
-        isGlobal: true,
-        isSignatureValidation: true,
-        isUserOpValidation: true,
-      },
-      selectors: [],
-      installData: "0x",
-      hooks: [
-        {
-          hookConfig: {
-            address: getDefaultTimeRangeModuleAddress(provider.chain),
-            entityId: 0, // uint32
-            hookType: HookType.VALIDATION,
-            hasPreHooks: true,
-            hasPostHooks: true,
-          },
-          initData: hookInstallData,
-        },
-      ],
-    });
-
-    // verify hook installtion succeeded
-    await provider.waitForUserOperationTransaction(installResult);
-
-    const hookUninstallData = TimeRangeModule.encodeOnUninstallData({
-      entityId: 0,
-    });
-
-    const uninstallResult = await provider.uninstallValidation({
-      moduleAddress: zeroAddress,
-      entityId: 0,
-      uninstallData: "0x",
-      hookUninstallDatas: [hookUninstallData],
-    });
-
-    // verify uninstall
-    await expect(
-      provider.waitForUserOperationTransaction(uninstallResult)
-    ).resolves.not.toThrowError();
-  });
-
-  it("installs time range module, then uninstalls module outside valid time range", async () => {
-    let provider = (
+    // create session key client
+    const sessionKeyProvider = (
       await givenConnectedProvider({
-        signer,
+        signer: sessionKey,
+        accountAddress: provider.account.address,
+        signerEntity: { entityId: 1, isGlobalValidation: true },
       })
     ).extend(installValidationActions);
 
-    await setBalance(client, {
-      address: provider.getAddress(),
-      value: parseEther("2"),
-    });
-
     const hookInstallData = TimeRangeModule.encodeOnInstallData({
-      entityId: 0,
-      validAfter: 0,
-      validUntil: 1,
+      entityId: 1,
+      validAfter: 1734507101,
+      validUntil: 1934507101,
     });
 
     const installResult = await provider.installValidation({
       validationConfig: {
-        moduleAddress: zeroAddress,
-        entityId: 0,
+        moduleAddress: getDefaultSingleSignerValidationModuleAddress(
+          provider.chain
+        ),
+        entityId: 1,
         isGlobal: true,
         isSignatureValidation: true,
         isUserOpValidation: true,
       },
       selectors: [],
-      installData: "0x",
+      installData: SingleSignerValidationModule.encodeOnInstallData({
+        entityId: 1,
+        signer: await sessionKey.getAddress(),
+      }),
       hooks: [
         {
           hookConfig: {
             address: getDefaultTimeRangeModuleAddress(provider.chain),
-            entityId: 0, // uint32
+            entityId: 1,
             hookType: HookType.VALIDATION,
             hasPreHooks: true,
-            hasPostHooks: true,
+            hasPostHooks: false,
           },
           initData: hookInstallData,
         },
@@ -831,38 +913,286 @@ describe("MA v2 Tests", async () => {
     // verify hook installation succeeded
     await provider.waitForUserOperationTransaction(installResult);
 
-    const hookUninstallData = TimeRangeModule.encodeOnUninstallData({
-      entityId: 0,
+    client.setAutomine(false);
+
+    // force block timestamp to be inside of range
+    await client.setNextBlockTimestamp({
+      timestamp: 1754507101n,
     });
 
-    const uninstallResult = await provider.uninstallValidation({
-      moduleAddress: zeroAddress,
-      entityId: 0,
-      uninstallData: "0x",
-      hookUninstallDatas: [hookUninstallData],
+    await client.mine({
+      blocks: 1,
     });
 
-    // uninstall should fail
-    await expect(
-      provider.waitForUserOperationTransaction(uninstallResult)
-    ).rejects.toThrowError();
+    const uo = await sessionKeyProvider.buildUserOperation({
+      uo: {
+        target,
+        data: "0x",
+      },
+    });
+
+    const signedUO = (await sessionKeyProvider.signUserOperation({
+      uoStruct: uo,
+    })) as UserOperationRequest_v7;
+
+    // calls entrypoint directly
+    await client.simulateContract({
+      address: sessionKeyProvider.account.getEntryPoint().address,
+      abi: entryPoint07Abi,
+      functionName: "handleOps",
+      args: [
+        [
+          {
+            sender: sessionKeyProvider.account.address,
+            nonce: fromHex(signedUO.nonce, "bigint"),
+            initCode:
+              signedUO.factory && signedUO.factoryData
+                ? concat([signedUO.factory, signedUO.factoryData])
+                : "0x",
+            callData: signedUO.callData,
+            accountGasLimits: packAccountGasLimits({
+              verificationGasLimit: signedUO.verificationGasLimit,
+              callGasLimit: signedUO.callGasLimit,
+            }),
+            preVerificationGas: fromHex(signedUO.preVerificationGas, "bigint"),
+            gasFees: packAccountGasLimits({
+              maxPriorityFeePerGas: signedUO.maxPriorityFeePerGas,
+              maxFeePerGas: signedUO.maxFeePerGas,
+            }),
+            paymasterAndData:
+              signedUO.paymaster && isAddress(signedUO.paymaster)
+                ? packPaymasterData({
+                    paymaster: signedUO.paymaster,
+                    paymasterVerificationGasLimit:
+                      signedUO.paymasterVerificationGasLimit,
+                    paymasterPostOpGasLimit: signedUO.paymasterPostOpGasLimit,
+                    paymasterData: signedUO.paymasterData,
+                  })
+                : "0x",
+            signature: signedUO.signature,
+          },
+        ],
+        provider.account.address,
+      ],
+      account: await sessionKeyProvider.account.getSigner().getAddress(),
+    });
+
+    client.setAutomine(true);
+  });
+
+  // NOTE: uses different validation and hook entity id than previous test because we do not uninstall the hook in the previous test
+  it("installs time range module, tries to send transaction outside valid time range", async () => {
+    let provider = (
+      await givenConnectedProvider({
+        signer,
+      })
+    ).extend(installValidationActions);
+
+    await setBalance(client, {
+      address: provider.getAddress(),
+      value: parseEther("2"),
+    });
+
+    // create session key client
+    const sessionKeyProvider = (
+      await givenConnectedProvider({
+        signer: sessionKey,
+        accountAddress: provider.account.address,
+        signerEntity: { entityId: 2, isGlobalValidation: true },
+      })
+    ).extend(installValidationActions);
+
+    const hookInstallData = TimeRangeModule.encodeOnInstallData({
+      entityId: 2,
+      validAfter: 1734507101,
+      validUntil: 1934507101,
+    });
+
+    const installResult = await provider.installValidation({
+      validationConfig: {
+        moduleAddress: getDefaultSingleSignerValidationModuleAddress(
+          provider.chain
+        ),
+        entityId: 2,
+        isGlobal: true,
+        isSignatureValidation: true,
+        isUserOpValidation: true,
+      },
+      selectors: [],
+      installData: SingleSignerValidationModule.encodeOnInstallData({
+        entityId: 2,
+        signer: await sessionKey.getAddress(),
+      }),
+      hooks: [
+        {
+          hookConfig: {
+            address: getDefaultTimeRangeModuleAddress(provider.chain),
+            entityId: 2,
+            hookType: HookType.VALIDATION,
+            hasPreHooks: true,
+            hasPostHooks: false,
+          },
+          initData: hookInstallData,
+        },
+      ],
+    });
+
+    // verify hook installation succeeded
+    await provider.waitForUserOperationTransaction(installResult);
+
+    client.setAutomine(false);
+
+    // force block timestamp to be outside of range
+    await client.setNextBlockTimestamp({
+      timestamp: 2054507101n,
+    });
+
+    await client.mine({
+      blocks: 1,
+    });
+
+    const uo = await sessionKeyProvider.buildUserOperation({
+      uo: {
+        target,
+        data: "0x",
+      },
+    });
+
+    const signedUO = (await sessionKeyProvider.signUserOperation({
+      uoStruct: uo,
+    })) as UserOperationRequest_v7;
+
+    // calls entrypoint directly
+    try {
+      await client.simulateContract({
+        address: sessionKeyProvider.account.getEntryPoint().address,
+        abi: entryPoint07Abi,
+        functionName: "handleOps",
+        args: [
+          [
+            {
+              sender: sessionKeyProvider.account.address,
+              nonce: fromHex(signedUO.nonce, "bigint"),
+              initCode:
+                signedUO.factory && signedUO.factoryData
+                  ? concat([signedUO.factory, signedUO.factoryData])
+                  : "0x",
+              callData: signedUO.callData,
+              accountGasLimits: packAccountGasLimits({
+                verificationGasLimit: signedUO.verificationGasLimit,
+                callGasLimit: signedUO.callGasLimit,
+              }),
+              preVerificationGas: fromHex(
+                signedUO.preVerificationGas,
+                "bigint"
+              ),
+              gasFees: packAccountGasLimits({
+                maxPriorityFeePerGas: signedUO.maxPriorityFeePerGas,
+                maxFeePerGas: signedUO.maxFeePerGas,
+              }),
+              paymasterAndData:
+                signedUO.paymaster && isAddress(signedUO.paymaster)
+                  ? packPaymasterData({
+                      paymaster: signedUO.paymaster,
+                      paymasterVerificationGasLimit:
+                        signedUO.paymasterVerificationGasLimit,
+                      paymasterPostOpGasLimit: signedUO.paymasterPostOpGasLimit,
+                      paymasterData: signedUO.paymasterData,
+                    })
+                  : "0x",
+              signature: signedUO.signature,
+            },
+          ],
+          provider.account.address,
+        ],
+        account: await sessionKeyProvider.account.getSigner().getAddress(),
+      });
+    } catch (err: any) {
+      // verify that simulation fails due to violation of time range restriction on session key
+      assert(
+        err.metaMessages.some((str: string) =>
+          str.includes("AA22 expired or not due")
+        )
+      );
+    }
+
+    client.setAutomine(true);
+  });
+
+  it("upgrade from a lightaccount", async () => {
+    const lightAccountClient = await createLightAccountClient({
+      chain: instance.chain,
+      signer,
+      transport: custom(instance.getClient()),
+      version: "v2.0.0",
+    });
+
+    await setBalance(client, {
+      address: lightAccountClient.getAddress(),
+      value: parseEther("2"),
+    });
+
+    const { createModularAccountV2FromExisting, ...upgradeToData } =
+      await getMAV2UpgradeToData(lightAccountClient, {
+        account: lightAccountClient.account,
+      });
+
+    await lightAccountClient.upgradeAccount({
+      upgradeTo: upgradeToData,
+      waitForTx: true,
+    });
+
+    const maV2Client = createSmartAccountClient({
+      chain: instance.chain,
+      transport: custom(client),
+      account: await createModularAccountV2FromExisting(),
+    });
+
+    // test uo
+
+    const startingAddressBalance = await getTargetBalance();
+
+    const result = await maV2Client.sendUserOperation({
+      uo: {
+        target: target,
+        value: sendAmount,
+        data: "0x",
+      },
+    });
+
+    await maV2Client.waitForUserOperationTransaction(result);
+
+    await expect(getTargetBalance()).resolves.toEqual(
+      startingAddressBalance + sendAmount
+    );
   });
 
   const givenConnectedProvider = async ({
     signer,
+    signerEntity,
     accountAddress,
-    usePaymaster = false,
+    paymasterMiddleware,
   }: {
     signer: SmartAccountSigner;
+    signerEntity?: SignerEntity;
     accountAddress?: `0x${string}`;
-    usePaymaster?: boolean;
+    paymasterMiddleware?: "alchemyGasAndPaymasterAndData" | "erc7677";
   }) =>
     createModularAccountV2Client({
       chain: instance.chain,
       signer,
       accountAddress,
+      signerEntity,
       transport: custom(instance.getClient()),
-      ...(usePaymaster ? erc7677Middleware() : {}),
+      ...(paymasterMiddleware === "alchemyGasAndPaymasterAndData"
+        ? alchemyGasAndPaymasterAndDataMiddleware({
+            policyId: "FAKE_POLICY_ID",
+            // @ts-ignore (expects an alchemy transport, but we're using a custom transport for mocking)
+            transport: custom(instance.getClient()),
+          })
+        : paymasterMiddleware === "erc7677"
+        ? erc7677Middleware()
+        : {}),
     });
 
   it("alchemy client calls the createAlchemySmartAccountClient", async () => {