+++ examples/ui-demo/src/components/small-cards/TransactionsCard.tsx
@@ -50,16 +50,16 @@ const TransactionsCardInner = ({
   }, [cardStatus]);
 
   return (
-    <div className="bg-bg-surface-default rounded-lg p-4 w-full xl:p-6 xl:w-[326px] xl:h-[478px] flex flex-col shadow-smallCard mb-5 xl:mb-0">
-      <div className="flex gap-3 xl:gap-0 xl:flex-col">
-        <div className="flex-shrink-0 bg-[#EAEBFE] rounded-xl mb-4 flex justify-center items-center relative h-[67px] w-[60px] sm:h-[154px] sm:w-[140px] xl:h-[222px] xl:w-full">
+    <div className="bg-bg-surface-default rounded-lg p-4 xl:p-6 w-full xl:w-[326px] xl:h-[500px] flex flex-col shadow-smallCard min-h-[220px]">
+      <div className="flex xl:flex-col gap-4">
+        <div className="flex-shrink-0 bg-[#EAEBFE] rounded-xl sm:mb-3 xl:mb-0 flex justify-center items-center relative h-[67px] w-[60px] sm:h-[154px] sm:w-[140px] xl:h-[222px] xl:w-full">
           <p className="absolute top-[-6px] left-[-6px] sm:top-1 sm:left-1 xl:left-auto xl:right-4 xl:top-4 px-2 py-1 font-semibold rounded-md text-xs text-[#7c3AED] bg-[#F3F3FF]">
             New!
           </p>
           <Key className="h-9 w-9 sm:h-[74px] sm:w-[74px] xl:h-[94px] xl:w-[94px]" />
         </div>
-        <div className="mb-3">
-          <h3 className="text-fg-primary  xl:text-xl font-semibold mb-2 xl:mb-3">
+        <div className="w-full">
+          <h3 className="text-fg-primary xl:text-xl font-semibold mb-2 xl:mb-3">
             Recurring transactions
           </h3>
 
@@ -71,9 +71,7 @@ const TransactionsCardInner = ({
           ) : cardStatus === "setup" ? (
             <div className="flex items-center">
               <LoadingIcon className="h-4 w-4 mr-2" />
-              <p className="text-fg-primary text-sm">
-                Creating session key and minting USDC...
-              </p>
+              <p className="text-fg-primary text-sm">Creating session key...</p>
             </div>
           ) : (
             <Transactions transactions={transactions} />
@@ -81,7 +79,7 @@ const TransactionsCardInner = ({
         </div>
       </div>
       <Button
-        className="mt-auto"
+        className="mt-auto w-full"
         onClick={handleTransactions}
         disabled={
           isLoadingClient || cardStatus === "setup" || cardStatus === "active"