+++ examples/ui-demo/src/components/small-cards/Wrapper.tsx
@@ -1,5 +1,4 @@
 import { useConfigStore } from "@/state";
-import { WalletTypes } from "@/app/config";
 import { MintCard7702, MintCardDefault } from "./MintCard";
 import {
   TransactionsCard7702,
@@ -7,11 +6,11 @@ import {
 } from "./TransactionsCard";
 
 export const SmallCardsWrapper = () => {
-  const { walletType } = useConfigStore();
+  const { accountMode } = useConfigStore();
 
   return (
-    <div className="flex flex-col xl:flex-row gap-6 lg:mt-6 items-center p-6">
-      {walletType === WalletTypes.smart ? (
+    <div className="flex flex-col xl:flex-row gap-6 lg:mt-6 items-center p-6 w-full justify-center max-w-screen-sm xl:max-w-none">
+      {accountMode === "default" ? (
         <>
           <MintCardDefault />
           <TransactionsCardDefault />